MENU - rozdzielenie funkcji do osobnych plików
