--- projekt/cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-78f8518a609fd265bc53.json
@@ -1,202 +1,0 @@
-{
-	"configurations" : 
-	[
-		{
-			"directories" : 
-			[
-				{
-					"build" : ".",
-					"childIndexes" : 
-					[
-						1
-					],
-					"hasInstallRule" : true,
-					"jsonFile" : "directory-.-Debug-d0094a50bb2071803777.json",
-					"minimumCMakeVersion" : 
-					{
-						"string" : "3.13"
-					},
-					"projectIndex" : 0,
-					"source" : ".",
-					"targetIndexes" : 
-					[
-						4,
-						5
-					]
-				},
-				{
-					"build" : "googletest-master",
-					"childIndexes" : 
-					[
-						2
-					],
-					"hasInstallRule" : true,
-					"jsonFile" : "directory-googletest-master-Debug-7aa0c1cf4c33ccc3044c.json",
-					"minimumCMakeVersion" : 
-					{
-						"string" : "3.13"
-					},
-					"parentIndex" : 0,
-					"projectIndex" : 1,
-					"source" : "googletest-master"
-				},
-				{
-					"build" : "googletest-master/googlemock",
-					"childIndexes" : 
-					[
-						3
-					],
-					"hasInstallRule" : true,
-					"jsonFile" : "directory-googletest-master.googlemock-Debug-2d83c415f49f4ab7737d.json",
-					"minimumCMakeVersion" : 
-					{
-						"string" : "3.13"
-					},
-					"parentIndex" : 1,
-					"projectIndex" : 2,
-					"source" : "googletest-master/googlemock",
-					"targetIndexes" : 
-					[
-						0,
-						1
-					]
-				},
-				{
-					"build" : "googletest-master/googletest",
-					"hasInstallRule" : true,
-					"jsonFile" : "directory-googletest-master.googletest-Debug-0a8a080b148a368a7503.json",
-					"minimumCMakeVersion" : 
-					{
-						"string" : "3.13"
-					},
-					"parentIndex" : 2,
-					"projectIndex" : 3,
-					"source" : "googletest-master/googletest",
-					"targetIndexes" : 
-					[
-						2,
-						3
-					]
-				}
-			],
-			"name" : "Debug",
-			"projects" : 
-			[
-				{
-					"childIndexes" : 
-					[
-						1
-					],
-					"directoryIndexes" : 
-					[
-						0
-					],
-					"name" : "projekt",
-					"targetIndexes" : 
-					[
-						4,
-						5
-					]
-				},
-				{
-					"childIndexes" : 
-					[
-						2
-					],
-					"directoryIndexes" : 
-					[
-						1
-					],
-					"name" : "googletest-distribution",
-					"parentIndex" : 0
-				},
-				{
-					"childIndexes" : 
-					[
-						3
-					],
-					"directoryIndexes" : 
-					[
-						2
-					],
-					"name" : "gmock",
-					"parentIndex" : 1,
-					"targetIndexes" : 
-					[
-						0,
-						1
-					]
-				},
-				{
-					"directoryIndexes" : 
-					[
-						3
-					],
-					"name" : "gtest",
-					"parentIndex" : 2,
-					"targetIndexes" : 
-					[
-						2,
-						3
-					]
-				}
-			],
-			"targets" : 
-			[
-				{
-					"directoryIndex" : 2,
-					"id" : "gmock::@007f8d9bf340d4fe6913",
-					"jsonFile" : "target-gmock-Debug-35b1d93def357e616a13.json",
-					"name" : "gmock",
-					"projectIndex" : 2
-				},
-				{
-					"directoryIndex" : 2,
-					"id" : "gmock_main::@007f8d9bf340d4fe6913",
-					"jsonFile" : "target-gmock_main-Debug-94fd5477f38993ee0418.json",
-					"name" : "gmock_main",
-					"projectIndex" : 2
-				},
-				{
-					"directoryIndex" : 3,
-					"id" : "gtest::@f5d8686f3b9b4fbaa54d",
-					"jsonFile" : "target-gtest-Debug-a1cd1165416ad905793a.json",
-					"name" : "gtest",
-					"projectIndex" : 3
-				},
-				{
-					"directoryIndex" : 3,
-					"id" : "gtest_main::@f5d8686f3b9b4fbaa54d",
-					"jsonFile" : "target-gtest_main-Debug-e6552d035d8b3c8057d9.json",
-					"name" : "gtest_main",
-					"projectIndex" : 3
-				},
-				{
-					"directoryIndex" : 0,
-					"id" : "projekt__debug::@6890427a1f51a3e7e1df",
-					"jsonFile" : "target-projekt__debug-Debug-90fba91b3c3f45e89782.json",
-					"name" : "projekt__debug",
-					"projectIndex" : 0
-				},
-				{
-					"directoryIndex" : 0,
-					"id" : "projekt__test::@6890427a1f51a3e7e1df",
-					"jsonFile" : "target-projekt__test-Debug-1b8603aac3f177508cfe.json",
-					"name" : "projekt__test",
-					"projectIndex" : 0
-				}
-			]
-		}
-	],
-	"kind" : "codemodel",
-	"paths" : 
-	{
-		"build" : "C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug",
-		"source" : "C:/Users/ASUS/Desktop/kalkulator/projekt"
-	},
-	"version" : 
-	{
-		"major" : 2,
-		"minor" : 6
-	}
-}
