--- projekt/cmake-build-debug/.cmake/api/v1/reply/target-gmock-Debug-35b1d93def357e616a13.json
@@ -1,206 +1,0 @@
-{
-	"archive" : {},
-	"artifacts" : 
-	[
-		{
-			"path" : "lib/libgmock.a"
-		}
-	],
-	"backtrace" : 3,
-	"backtraceGraph" : 
-	{
-		"commands" : 
-		[
-			"add_library",
-			"cxx_library_with_type",
-			"cxx_library",
-			"install",
-			"install_project",
-			"target_link_libraries",
-			"include_directories",
-			"include",
-			"target_compile_features"
-		],
-		"files" : 
-		[
-			"googletest-master/googletest/cmake/internal_utils.cmake",
-			"googletest-master/googlemock/CMakeLists.txt",
-			"SetCommonConfig.cmake",
-			"CMakeLists.txt"
-		],
-		"nodes" : 
-		[
-			{
-				"file" : 1
-			},
-			{
-				"command" : 2,
-				"file" : 1,
-				"line" : 96,
-				"parent" : 0
-			},
-			{
-				"command" : 1,
-				"file" : 0,
-				"line" : 210,
-				"parent" : 1
-			},
-			{
-				"command" : 0,
-				"file" : 0,
-				"line" : 165,
-				"parent" : 2
-			},
-			{
-				"command" : 4,
-				"file" : 1,
-				"line" : 115,
-				"parent" : 0
-			},
-			{
-				"command" : 3,
-				"file" : 0,
-				"line" : 303,
-				"parent" : 4
-			},
-			{
-				"command" : 5,
-				"file" : 1,
-				"line" : 97,
-				"parent" : 0
-			},
-			{
-				"file" : 3
-			},
-			{
-				"command" : 7,
-				"file" : 3,
-				"line" : 18,
-				"parent" : 7
-			},
-			{
-				"file" : 2,
-				"parent" : 8
-			},
-			{
-				"command" : 6,
-				"file" : 2,
-				"line" : 29,
-				"parent" : 9
-			},
-			{
-				"command" : 6,
-				"file" : 1,
-				"line" : 74,
-				"parent" : 0
-			},
-			{
-				"command" : 8,
-				"file" : 0,
-				"line" : 198,
-				"parent" : 2
-			}
-		]
-	},
-	"compileGroups" : 
-	[
-		{
-			"compileCommandFragments" : 
-			[
-				{
-					"fragment" : "-g -std=c++17 -fdiagnostics-color=always"
-				},
-				{
-					"fragment" : "-Wall -Wshadow -Wundef -Wno-error=dangling-else -DGTEST_HAS_PTHREAD=0 -fexceptions -Wextra -Wno-unused-parameter -Wno-missing-field-initializers"
-				}
-			],
-			"includes" : 
-			[
-				{
-					"backtrace" : 10,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/include"
-				},
-				{
-					"backtrace" : 11,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googlemock/include"
-				},
-				{
-					"backtrace" : 11,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googlemock"
-				},
-				{
-					"backtrace" : 11,
-					"isSystem" : true,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googletest/include"
-				},
-				{
-					"backtrace" : 11,
-					"isSystem" : true,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googletest"
-				}
-			],
-			"language" : "CXX",
-			"languageStandard" : 
-			{
-				"backtraces" : 
-				[
-					12
-				],
-				"standard" : "17"
-			},
-			"sourceIndexes" : 
-			[
-				0
-			]
-		}
-	],
-	"dependencies" : 
-	[
-		{
-			"backtrace" : 6,
-			"id" : "gtest::@f5d8686f3b9b4fbaa54d"
-		}
-	],
-	"id" : "gmock::@007f8d9bf340d4fe6913",
-	"install" : 
-	{
-		"destinations" : 
-		[
-			{
-				"backtrace" : 5,
-				"path" : "lib"
-			}
-		],
-		"prefix" : 
-		{
-			"path" : "C:/Program Files (x86)/Project"
-		}
-	},
-	"name" : "gmock",
-	"nameOnDisk" : "libgmock.a",
-	"paths" : 
-	{
-		"build" : "googletest-master/googlemock",
-		"source" : "googletest-master/googlemock"
-	},
-	"sourceGroups" : 
-	[
-		{
-			"name" : "Source Files",
-			"sourceIndexes" : 
-			[
-				0
-			]
-		}
-	],
-	"sources" : 
-	[
-		{
-			"backtrace" : 3,
-			"compileGroupIndex" : 0,
-			"path" : "googletest-master/googlemock/src/gmock-all.cc",
-			"sourceGroupIndex" : 0
-		}
-	],
-	"type" : "STATIC_LIBRARY"
-}
