--- projekt/cmake-build-debug/.cmake/api/v1/reply/target-gtest-Debug-a1cd1165416ad905793a.json
@@ -1,182 +1,0 @@
-{
-	"archive" : {},
-	"artifacts" : 
-	[
-		{
-			"path" : "lib/libgtest.a"
-		}
-	],
-	"backtrace" : 3,
-	"backtraceGraph" : 
-	{
-		"commands" : 
-		[
-			"add_library",
-			"cxx_library_with_type",
-			"cxx_library",
-			"install",
-			"install_project",
-			"include_directories",
-			"include",
-			"target_compile_features"
-		],
-		"files" : 
-		[
-			"googletest-master/googletest/cmake/internal_utils.cmake",
-			"googletest-master/googletest/CMakeLists.txt",
-			"SetCommonConfig.cmake",
-			"CMakeLists.txt"
-		],
-		"nodes" : 
-		[
-			{
-				"file" : 1
-			},
-			{
-				"command" : 2,
-				"file" : 1,
-				"line" : 123,
-				"parent" : 0
-			},
-			{
-				"command" : 1,
-				"file" : 0,
-				"line" : 210,
-				"parent" : 1
-			},
-			{
-				"command" : 0,
-				"file" : 0,
-				"line" : 165,
-				"parent" : 2
-			},
-			{
-				"command" : 4,
-				"file" : 1,
-				"line" : 158,
-				"parent" : 0
-			},
-			{
-				"command" : 3,
-				"file" : 0,
-				"line" : 303,
-				"parent" : 4
-			},
-			{
-				"file" : 3
-			},
-			{
-				"command" : 6,
-				"file" : 3,
-				"line" : 18,
-				"parent" : 6
-			},
-			{
-				"file" : 2,
-				"parent" : 7
-			},
-			{
-				"command" : 5,
-				"file" : 2,
-				"line" : 29,
-				"parent" : 8
-			},
-			{
-				"command" : 5,
-				"file" : 1,
-				"line" : 113,
-				"parent" : 0
-			},
-			{
-				"command" : 7,
-				"file" : 0,
-				"line" : 198,
-				"parent" : 2
-			}
-		]
-	},
-	"compileGroups" : 
-	[
-		{
-			"compileCommandFragments" : 
-			[
-				{
-					"fragment" : "-g -std=c++17 -fdiagnostics-color=always"
-				},
-				{
-					"fragment" : "-Wall -Wshadow -Wundef -Wno-error=dangling-else -DGTEST_HAS_PTHREAD=0 -fexceptions -Wextra -Wno-unused-parameter -Wno-missing-field-initializers"
-				}
-			],
-			"includes" : 
-			[
-				{
-					"backtrace" : 9,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/include"
-				},
-				{
-					"backtrace" : 10,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googletest/include"
-				},
-				{
-					"backtrace" : 10,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googletest"
-				}
-			],
-			"language" : "CXX",
-			"languageStandard" : 
-			{
-				"backtraces" : 
-				[
-					11
-				],
-				"standard" : "17"
-			},
-			"sourceIndexes" : 
-			[
-				0
-			]
-		}
-	],
-	"id" : "gtest::@f5d8686f3b9b4fbaa54d",
-	"install" : 
-	{
-		"destinations" : 
-		[
-			{
-				"backtrace" : 5,
-				"path" : "lib"
-			}
-		],
-		"prefix" : 
-		{
-			"path" : "C:/Program Files (x86)/Project"
-		}
-	},
-	"name" : "gtest",
-	"nameOnDisk" : "libgtest.a",
-	"paths" : 
-	{
-		"build" : "googletest-master/googletest",
-		"source" : "googletest-master/googletest"
-	},
-	"sourceGroups" : 
-	[
-		{
-			"name" : "Source Files",
-			"sourceIndexes" : 
-			[
-				0
-			]
-		}
-	],
-	"sources" : 
-	[
-		{
-			"backtrace" : 3,
-			"compileGroupIndex" : 0,
-			"path" : "googletest-master/googletest/src/gtest-all.cc",
-			"sourceGroupIndex" : 0
-		}
-	],
-	"type" : "STATIC_LIBRARY"
-}
