--- projekt/cmake-build-debug/.cmake/api/v1/reply/target-projekt__test-Debug-1b8603aac3f177508cfe.json
@@ -1,237 +1,0 @@
-{
-	"artifacts" : 
-	[
-		{
-			"path" : "projekt__test.exe"
-		},
-		{
-			"path" : "projekt__test.pdb"
-		}
-	],
-	"backtrace" : 3,
-	"backtraceGraph" : 
-	{
-		"commands" : 
-		[
-			"add_executable",
-			"include",
-			"target_link_libraries",
-			"include_directories",
-			"target_include_directories"
-		],
-		"files" : 
-		[
-			"SetCommonConfig.cmake",
-			"CMakeLists.txt",
-			"googletest-master/googlemock/CMakeLists.txt"
-		],
-		"nodes" : 
-		[
-			{
-				"file" : 1
-			},
-			{
-				"command" : 1,
-				"file" : 1,
-				"line" : 18,
-				"parent" : 0
-			},
-			{
-				"file" : 0,
-				"parent" : 1
-			},
-			{
-				"command" : 0,
-				"file" : 0,
-				"line" : 48,
-				"parent" : 2
-			},
-			{
-				"command" : 2,
-				"file" : 0,
-				"line" : 69,
-				"parent" : 2
-			},
-			{
-				"file" : 2
-			},
-			{
-				"command" : 2,
-				"file" : 2,
-				"line" : 97,
-				"parent" : 5
-			},
-			{
-				"command" : 3,
-				"file" : 0,
-				"line" : 29,
-				"parent" : 2
-			},
-			{
-				"command" : 4,
-				"file" : 0,
-				"line" : 61,
-				"parent" : 2
-			}
-		]
-	},
-	"compileGroups" : 
-	[
-		{
-			"compileCommandFragments" : 
-			[
-				{
-					"fragment" : "-g -std=gnu++17 -fdiagnostics-color=always"
-				}
-			],
-			"includes" : 
-			[
-				{
-					"backtrace" : 7,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/include"
-				},
-				{
-					"backtrace" : 8,
-					"isSystem" : true,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googlemock/include"
-				},
-				{
-					"backtrace" : 8,
-					"isSystem" : true,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googletest/include"
-				},
-				{
-					"backtrace" : 4,
-					"isSystem" : true,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googlemock"
-				},
-				{
-					"backtrace" : 4,
-					"isSystem" : true,
-					"path" : "C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googletest"
-				}
-			],
-			"language" : "CXX",
-			"languageStandard" : 
-			{
-				"backtraces" : 
-				[
-					4
-				],
-				"standard" : "17"
-			},
-			"sourceIndexes" : 
-			[
-				0,
-				1,
-				2,
-				3,
-				4,
-				5
-			]
-		}
-	],
-	"dependencies" : 
-	[
-		{
-			"backtrace" : 4,
-			"id" : "gtest::@f5d8686f3b9b4fbaa54d"
-		},
-		{
-			"backtrace" : 4,
-			"id" : "gmock::@007f8d9bf340d4fe6913"
-		}
-	],
-	"id" : "projekt__test::@6890427a1f51a3e7e1df",
-	"link" : 
-	{
-		"commandFragments" : 
-		[
-			{
-				"fragment" : "-g",
-				"role" : "flags"
-			},
-			{
-				"fragment" : "",
-				"role" : "flags"
-			},
-			{
-				"backtrace" : 4,
-				"fragment" : "lib\\libgmock.a",
-				"role" : "libraries"
-			},
-			{
-				"backtrace" : 6,
-				"fragment" : "lib\\libgtest.a",
-				"role" : "libraries"
-			},
-			{
-				"fragment" : "-lkernel32 -luser32 -lgdi32 -lwinspool -lshell32 -lole32 -loleaut32 -luuid -lcomdlg32 -ladvapi32",
-				"role" : "libraries"
-			}
-		],
-		"language" : "CXX"
-	},
-	"name" : "projekt__test",
-	"nameOnDisk" : "projekt__test.exe",
-	"paths" : 
-	{
-		"build" : ".",
-		"source" : "."
-	},
-	"sourceGroups" : 
-	[
-		{
-			"name" : "Source Files",
-			"sourceIndexes" : 
-			[
-				0,
-				1,
-				2,
-				3,
-				4,
-				5
-			]
-		}
-	],
-	"sources" : 
-	[
-		{
-			"backtrace" : 3,
-			"compileGroupIndex" : 0,
-			"path" : "src/projekt.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 3,
-			"compileGroupIndex" : 0,
-			"path" : "test/test_kalkulator_prosty.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 3,
-			"compileGroupIndex" : 0,
-			"path" : "test/test_zespolone.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 3,
-			"compileGroupIndex" : 0,
-			"path" : "test/test_kalkulator_macierzy.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 3,
-			"compileGroupIndex" : 0,
-			"path" : "test/test_pola.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 3,
-			"compileGroupIndex" : 0,
-			"path" : "test/main_test.cpp",
-			"sourceGroupIndex" : 0
-		}
-	],
-	"type" : "EXECUTABLE"
-}
