+++ projekt/cmake-build-debug/CMakeFiles/3.27.8/CMakeSystem.cmake
@@ -1,13 +1,13 @@
-set(CMAKE_HOST_SYSTEM "Windows-10.0.19045")
+set(CMAKE_HOST_SYSTEM "Windows-10.0.22631")
 set(CMAKE_HOST_SYSTEM_NAME "Windows")
-set(CMAKE_HOST_SYSTEM_VERSION "10.0.19045")
+set(CMAKE_HOST_SYSTEM_VERSION "10.0.22631")
 set(CMAKE_HOST_SYSTEM_PROCESSOR "AMD64")
 
 
 
-set(CMAKE_SYSTEM "Windows-10.0.19045")
+set(CMAKE_SYSTEM "Windows-10.0.22631")
 set(CMAKE_SYSTEM_NAME "Windows")
-set(CMAKE_SYSTEM_VERSION "10.0.19045")
+set(CMAKE_SYSTEM_VERSION "10.0.22631")
 set(CMAKE_SYSTEM_PROCESSOR "AMD64")
 
 set(CMAKE_CROSSCOMPILING "FALSE")
