+++ projekt/cmake-build-debug/CMakeFiles/TargetDirectories.txt
@@ -1,30 +1,30 @@
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/CMakeFiles/projekt__debug.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/CMakeFiles/projekt__test.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/CMakeFiles/edit_cache.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/CMakeFiles/rebuild_cache.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/CMakeFiles/list_install_components.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/CMakeFiles/install.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/CMakeFiles/install/local.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/CMakeFiles/install/strip.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/edit_cache.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/rebuild_cache.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/list_install_components.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/install.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/install/local.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/install/strip.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/gmock.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/gmock_main.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/edit_cache.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/rebuild_cache.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/list_install_components.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/install.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/install/local.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/install/strip.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/gtest.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/gtest_main.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/edit_cache.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/rebuild_cache.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/list_install_components.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/install.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/install/local.dir
-C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/install/strip.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/CMakeFiles/projekt__debug.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/CMakeFiles/projekt__test.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/CMakeFiles/edit_cache.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/CMakeFiles/rebuild_cache.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/CMakeFiles/list_install_components.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/CMakeFiles/install.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/CMakeFiles/install/local.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/CMakeFiles/install/strip.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/edit_cache.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/rebuild_cache.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/list_install_components.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/install.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/install/local.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/CMakeFiles/install/strip.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/gmock.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/gmock_main.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/edit_cache.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/rebuild_cache.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/list_install_components.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/install.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/install/local.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/CMakeFiles/install/strip.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/gtest.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/gtest_main.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/edit_cache.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/rebuild_cache.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/list_install_components.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/install.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/install/local.dir
+C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googletest/CMakeFiles/install/strip.dir
