+++ projekt/cmake-build-debug/CMakeFiles/clion-Debug-log.txt
@@ -1,7 +1,15 @@
-"C:\Program Files\JetBrains\CLion 2023.3.4\bin\cmake\win\x64\bin\cmake.exe" -DCMAKE_BUILD_TYPE=Debug "-DCMAKE_MAKE_PROGRAM=C:/Program Files/JetBrains/CLion 2023.3.4/bin/ninja/win/x64/ninja.exe" -G Ninja -S C:\Users\ASUS\Desktop\kalkulator\projekt -B C:\Users\ASUS\Desktop\kalkulator\projekt\cmake-build-debug
+"C:\Program Files\JetBrains\CLion 2023.3.4\bin\cmake\win\x64\bin\cmake.exe" -DCMAKE_BUILD_TYPE=Debug -G "CodeBlocks - MinGW Makefiles" -S C:\Users\user\Desktop\npg_kalkulator\projekt -B C:\Users\user\Desktop\npg_kalkulator\projekt\cmake-build-debug
+CMake Deprecation Warning:
+  Support for "Extra Generators" like
+
+    CodeBlocks
+
+  is deprecated and will be removed from a future version of CMake.  IDEs may
+  use the cmake-file-api(7) to view CMake-generated project build trees.
+
+
 -- PROJECT: projekt
 -- Using AddressSanitizer (ASan).
--- Could NOT find Python3 (missing: Python3_EXECUTABLE Interpreter) 
--- Configuring done (8.4s)
+-- Configuring done (0.7s)
 -- Generating done (0.2s)
--- Build files have been written to: C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug
+-- Build files have been written to: C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug
