+++ projekt/cmake-build-debug/CMakeFiles/clion-environment.txt
@@ -1,4 +1,4 @@
 ToolSet: 11.0 w64 (local)@C:\Program Files\JetBrains\CLion 2023.3.4\bin\mingw
 Options: 
 
-Options:-DCMAKE_MAKE_PROGRAM=C:/Program Files/JetBrains/CLion 2023.3.4/bin/ninja/win/x64/ninja.exe
+Options:
