+++ projekt/cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: May 31 15:48 Œrodkowoeuropejski czas letni
+Start testing: May 31 18:56 Œrodkowoeuropejski czas letni
 ----------------------------------------------------------
-End testing: May 31 15:48 Œrodkowoeuropejski czas letni
+End testing: May 31 18:56 Œrodkowoeuropejski czas letni
