+++ projekt/cmake-build-debug/googletest-master/cmake_install.cmake
@@ -1,8 +1,8 @@
-# Install script for directory: C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master
+# Install script for directory: C:/Users/user/Desktop/npg_kalkulator/projekt/googletest-master
 
 # Set the install prefix
 if(NOT DEFINED CMAKE_INSTALL_PREFIX)
-  set(CMAKE_INSTALL_PREFIX "C:/Program Files (x86)/Project")
+  set(CMAKE_INSTALL_PREFIX "C:/Program Files (x86)/projekt")
 endif()
 string(REGEX REPLACE "/$" "" CMAKE_INSTALL_PREFIX "${CMAKE_INSTALL_PREFIX}")
 
@@ -39,7 +39,7 @@ endif()
 
 if(NOT CMAKE_INSTALL_LOCAL_ONLY)
   # Include the install script for each subdirectory.
-  include("C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/cmake_install.cmake")
+  include("C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock/cmake_install.cmake")
 
 endif()
