+++ projekt/cmake-build-debug/googletest-master/googlemock/CTestTestfile.cmake
@@ -1,6 +1,6 @@
 # CMake generated Testfile for 
-# Source directory: C:/Users/ASUS/Desktop/kalkulator/projekt/googletest-master/googlemock
-# Build directory: C:/Users/ASUS/Desktop/kalkulator/projekt/cmake-build-debug/googletest-master/googlemock
+# Source directory: C:/Users/user/Desktop/npg_kalkulator/projekt/googletest-master/googlemock
+# Build directory: C:/Users/user/Desktop/npg_kalkulator/projekt/cmake-build-debug/googletest-master/googlemock
 # 
 # This file includes the relevant testing commands required for 
 # testing this directory and lists subdirectories to be tested as well.
