+++ projekt/main.cpp
@@ -4,21 +4,17 @@
 #include <vector>
 #include <stdexcept>
 
-// Jesli ktos to czyta to tlumacze te 4 bledy co sa
-// dwa z nich pojawiaja sie przez puste swiche i nimi sie wgl nie przejmujemy
-// dwa pozostale sa bo uzywamy comon configa z PSiO i nie podoba mu sie ze porownujemy cos do NULLa dla charow
-// wiec mozna sprobowac go jakos ladnie upakowac zeby sie nie czepialo ale raczej nic to nie zepsuje
-void czyszczenie(){
+void clear(){
     for (int i = 0; i < 25; i++) {
         std::cout << std::endl;
     }
 }
 
-void blad() {
+void mistake() {
     std::cout << "Nie znana komenda \nsprobuj jeszcze raz" << std::endl;
 }
 
-int repet() {
+int repeat() {
     std::string con;
     std::cout << "Chcesz kontynuowac? [T/N]" << std::endl;
     std::cin >> con;
@@ -28,16 +24,16 @@ int repet() {
 int main() {
 
     start:
-    czyszczenie();
+    clear();
     std::cout << "<<Witaj w kalkulatorze>>\n\nAby przejsc dalej wpisz NEXT" << std::endl;
     std::string t;
     std::cin >> t;
     if (t != "NEXT" && t != "Next" && t != "next") {
-        blad();
+        mistake();
         goto start;
     }
 
-    czyszczenie();
+    clear();
     p_menu:
     std::cout << "Jakiej funkcji chcesz uzyc:\n"
                  " 0. Help\n"
@@ -58,16 +54,16 @@ int main() {
     int fun;
     std::cin >> fun;
 
-    czyszczenie();
+    clear();
     switch (fun) {
         case 1:
-            prosty();
+            simple();
             goto p_menu;
         case 2:
             matrix();
             goto p_menu;
         case 3:
-            kalk_poteg();
+            extraction_exponentiation();
             goto p_menu;
         case 4:
             // Przerzut na przelicznik
@@ -76,7 +72,7 @@ int main() {
             // Przerzut na porownania
             goto p_menu;
         case 6:
-            print_zao();
+            rounding();
             goto p_menu;
         case 7: {
             char choice;
@@ -102,7 +98,7 @@ int main() {
                     std::cerr << e.what() << std::endl;
                 }
             } else {
-                blad();
+                mistake();
             }
             goto p_menu;
         }
@@ -110,7 +106,7 @@ int main() {
             area();
             goto p_menu;
         case 9:
-            logarytmy_by_Natalia();
+            logarithm_user();
             goto p_menu;
         case 10:
             cykl_o_tryg();
@@ -126,7 +122,7 @@ int main() {
         case 13:
             goto start;
         default:
-            blad();
+            mistake();
             goto p_menu;
     }
 
@@ -153,7 +149,7 @@ int main() {
 
     switch (help_fun) {
         case 1:
-            std::cout << "Kalkulator prosty, ktory obsluguje: dodawanie, odejmowanie, mnozenie, dzielenie, podnoszenie do zadanej potegi oraz obliczanie reszty z dzielenia." << std::endl;
+            std::cout << "Kalkulator prosty, ktory obsluguje: dodawanie, odejmowanie, mnozenie, dzielenie oraz obliczanie reszty z dzielenia." << std::endl;
             break;
         case 2:
             std::cout << "Program dodaje, odejmuje i mnozy macierze\nNa poczatku nalezy wybrac znak, pozniej wprowadzic odpowiednie macierze\n\n"
@@ -197,7 +193,7 @@ int main() {
         case 14:
             goto p_menu;
         default:
-            blad();
+            mistake();
             goto p_help;
     }
     goto p_menu;
