--- projekt/src/projekt.cpp
@@ -1,1026 +1,0 @@
-//
-// Created by Marek Generowicz on 27/04/2024.
-//
-#include "projekt.hpp"
-
-double potegowanie (double liczba, int potega){
-    double wynik = 1.0;
-    if (potega < 0){
-        for (int i = 0; - potega > i; i++){
-            wynik /= liczba;
-        }
-    }
-    else if (potega == 0){
-        wynik = 1.0;
-    }
-    else {
-        for (int i = 0; potega > i; i++) {
-            wynik *= liczba;
-        }
-    }
-return wynik;
-}
-
-double pierwiastkowanie(double liczba, int stopien){
-    if (liczba < 0 && (stopien % 2) == 0 ) {
-        return NAN;
-    }
-
-    double wynik = liczba / 2.0;
-    double przyblizenie = 0.00001;
-
-    while (std::abs(potegowanie(wynik, stopien) - liczba) > przyblizenie) {
-    wynik = ((stopien - 1.0) * wynik + liczba / potegowanie(wynik, stopien - 1.0)) / stopien;
-    }
-
-    wynik = std::round(wynik * 10000.0) / 10000.0;
-
-    return wynik;
-}
-
-void kalk_poteg (){
-    kalk_poteg:
-    std::cout <<"Wprowadz co chcesz zrobić\n* - potegowanie\ns - pierwiastkowanie" << std::endl;
-
-    char znak;
-    std::cin >> znak;
-    switch(znak){
-        case '*':
-            std::cout << "Podaj wartość jaką chcesz podnieść do potęgi" << std::endl;
-            double liczba;
-            std::cin >> liczba;
-            std::cout << "Podaj stopień do jakiego chcesz podnieść wcześniejszą wartość" << std::endl;
-            int stopien;
-            std::cin >> stopien;
-            std::cout << potegowanie(liczba, stopien) << std::endl;
-            break;
-
-        case 's':
-            std::cout << "Podaj wartość jaką chcesz pierwiastkowac" << std::endl;
-            double wartosc;
-            std::cin >> wartosc;
-            std::cout << "Podaj stopień piewiastka" << std::endl;
-            int pierw;
-            std::cin >> pierw;
-            std::cout << pierwiastkowanie(wartosc, pierw) << std::endl;
-            break;
-
-        default:
-            std::cout << "Nie znana komenda \nspróbuj jeszcze raz" << std::endl;
-            goto kalk_poteg;
-    }
-}
-
-double addition(double num1, double num2) {
-    return num1 + num2;
-}
-
-double subtraction(double num1, double num2) {
-    return num1 - num2;
-}
-
-double division(double num1, double num2) {
-    return num1 / num2;
-}
-
-double multiplication(double num1, double num2) {
-    return num1 * num2;
-}
-
-int modulo(int num1, int num2) {
-    return num1 % num2;
-}
-
-//funkcja tworzaca macierz o okreslonym rozmarze
-std::vector<std::vector<int>> createMatrix(std::size_t rows, std::size_t col) {
-    std::vector<std::vector<int>> matrix(rows, std::vector<int>(col));
-
-    for(std::size_t i = 0; i < rows; i++)
-        for(std::size_t j = 0; j < col; j++) {
-            std::cout << "[ " << i+1 << ", " << j+1 << "]: ";
-            std::cin >> matrix[i][j];
-        }
-
-    return matrix;
-}
-
-//funkcja pokazujaca na ekran macierz wynikowa
-void showMatrix(std::vector<std::vector<int>> matrix, std::size_t rows, std::size_t col) {
-    std::cout << "[" << std::endl;
-    for(std::size_t i = 0; i < rows; i++) {
-        for (std::size_t j = 0; j < col; j++)
-            std::cout << std::setw(3) << matrix[i][j] << " ";
-        std::cout << std::endl;
-    }
-    std::cout << "]" << std::endl;
-}
-
-//funkcja dodajaca macierze
-std::vector<std::vector<int>> addMatrix (std::vector<std::vector<int>> matrix1, std::vector<std::vector<int>> matrix2, std::size_t rows, std::size_t col){
-    std::vector<std::vector<int>> new_matrix (rows, std::vector<int>(col));
-    for(std::size_t i = 0; i < rows; i++)
-        for(std::size_t j = 0; j < col; j++)
-            new_matrix[i][j] = matrix1[i][j] + matrix2[i][j];
-    return new_matrix;
-}
-
-void addMatrix_user(){
-    std::size_t rows;
-    std::size_t col;
-    std::cout <<"Podaj rozmiar wierszy:\n";
-    std::cin >> rows;
-    std::cout <<"Podaj rozmiar kolumn:\n";
-    std::cin >> col;
-
-    std::cout <<"\nWpisywanie wartosci do pierwszej macierzy:\n";
-    std::vector<std::vector<int>> matrix1 = createMatrix(rows, col);
-    std::cout <<"\nWpisywanie wartosci do drugiej macierzy:\n";
-    std::vector<std::vector<int>> matrix2 = createMatrix(rows, col);
-
-    std::vector<std::vector<int>> new_matrix (rows, std::vector<int>(col));
-    new_matrix = addMatrix(matrix1, matrix2, rows, col);
-
-    std::cout << "\nWynik wykonanego dzialania:\n";
-    showMatrix(new_matrix ,rows ,col);
-}
-
-//funkcja odejmujaca macierze
-std::vector<std::vector<int>> substractMatrix (std::vector<std::vector<int>> matrix1, std::vector<std::vector<int>> matrix2, std::size_t rows, std::size_t col){
-    std::vector<std::vector<int>> new_matrix (rows, std::vector<int>(col));
-    for(std::size_t i = 0; i < rows; i++)
-        for(std::size_t j = 0; j < col; j++)
-            new_matrix[i][j] = matrix1[i][j] - matrix2[i][j];
-    return new_matrix;
-}
-
-void substractMatrix_user(){
-    std::size_t rows;
-    std::size_t col;
-    std::cout <<"Podaj rozmiar wierszy:\n";
-    std::cin >> rows;
-    std::cout <<"Podaj rozmiar kolumn:\n";
-    std::cin >> col;
-    std::cout << "Pamietaj zeby podac macierze w odpowiedniej kolejnosci, odejmowanie nie jest dzialaniem przemiennym" << std::endl;
-
-    std::cout <<"\nWpisywanie wartosci do pierwszej macierzy:\n";
-    std::vector<std::vector<int>> matrix1 = createMatrix(rows, col);
-    std::cout <<"\nWpisywanie wartosci do drugiej macierzy:\n";
-    std::vector<std::vector<int>> matrix2 = createMatrix(rows, col);
-
-    std::vector<std::vector<int>> new_matrix (rows, std::vector<int>(col));
-    new_matrix = substractMatrix(matrix1, matrix2, rows, col);
-
-    std::cout << "\nWynik wykonanego dzialania:\n";
-    showMatrix(new_matrix ,rows ,col);
-}
-
-//funkcja mnozaca macierze
-std::vector<std::vector<int>> multiMatrix(std::vector<std::vector<int>> matrix1, std::vector<std::vector<int>> matrix2, std::size_t rows1, std::size_t rows2, std::size_t col2){
-    std::vector<std::vector<int>> new_matrix (rows1, std::vector<int>(col2));
-    for(std::size_t i = 0; i < rows1; i++){
-        for(std::size_t j = 0; j < col2; j++){
-            new_matrix[i][j] = 0;
-            for(std::size_t k = 0; k < rows2; k++){
-                new_matrix[i][j] += matrix1[i][k] * matrix2[k][j];
-            }
-        }
-    }
-    return new_matrix;
-}
-
-void multiMatrix_user(){
-    std::cout << "Pamietaj ze rozmiar wierszy pierwszej macierzy = rozmiar kolumn drugiej macierzy\n oraz rozmiar kolumn pierwszej macierzy = rozmiar wierszy drugiej macierzy" << std::endl;
-    std::size_t rows1;
-    std::size_t col1;
-    std::size_t rows2;
-    std::size_t col2;
-    std::cout <<"Podaj rozmiar wierszy pierwszej macierzy:\n";
-    std::cin >> rows1;
-    col2 = rows1;
-    std::cout <<"Podaj rozmiar kolumn pierwszej macierzy:\n";
-    std::cin >> col1;
-    rows2 = col1;
-
-    std::cout <<"\nWpisywanie wartosci do pierwszej macierzy:\n";
-    std::vector<std::vector<int>> matrix1 = createMatrix(rows1, col1);
-    std::cout <<"\nWpisywanie wartosci do drugiej macierzy:\n";
-    std::vector<std::vector<int>> matrix2 = createMatrix(rows2, col2);
-
-    std::vector<std::vector<int>> new_matrix (rows1, std::vector<int>(col2));
-    new_matrix = multiMatrix(matrix1, matrix2, rows1, rows2, col2);
-
-    std::cout << "\nWynik wykonanego dzialania:\n";
-    showMatrix(new_matrix ,rows1 ,col2);
-}
-
-//wyznaczniki macierzy
-//funkcja sprawdzająca czy macierz jest kwadratowa
-bool sq_matrix(std::vector<std::vector<int>> Matrix){
-    if (Matrix.size() == Matrix[0].size()) {
-        return true;
-    }else {
-        return false;
-    }}
-
-//funkcja zmniejszająca macierz
-std::vector<std::vector<int>> Matrix_cutter(const std::vector<std::vector<int>>& Matrix, std::size_t row_to_cut, std::size_t col_to_cut){
-    std::vector<std::vector<int>> Cutted_Matrix;
-    for(std::size_t i = 0; i < Matrix.size(); i++){
-        if (i != row_to_cut - 1){
-            std::vector<int> row;
-            for (std::size_t j = 0; j < Matrix[0].size(); j++){
-                if (j != col_to_cut - 1){
-                    row.push_back(Matrix[i][j]);
-                }
-            }
-            Cutted_Matrix.push_back(row);
-        }
-    }
-    return Cutted_Matrix;
-}
-
-//funckaj licząca wartość wyznacznika
-int determinant(const std::vector<std::vector<int>>& Matrix){
-    if (Matrix.size() == 1){
-        return Matrix[0][0];
-    } else if (Matrix.size() == 2) {
-        return Matrix[0][0] * Matrix[1][1] - Matrix[0][1] * Matrix[1][0];
-    } else {
-        int deter = 0;
-        for (std::size_t i = 0; i < Matrix.size(); ++i) {
-            std::vector<std::vector<int>> subMatrix = Matrix_cutter(Matrix, 1, i+1);
-            deter += (i % 2 == 0 ? 1 : -1) * Matrix[0][i] * determinant(subMatrix);
-        }
-        return deter;
-    }
-}
-
-//funkcja do opisu do wyznacznika
-void determinat_str(){
-    std::size_t  r;
-    std::size_t  k;
-    std::cout << "Podaj liczbę wierszy i kolumn" << std::endl;
-    std::cin >> r;
-    std::cin >> k;
-    std::vector<std::vector<int>> Matrix = createMatrix(r,k);
-    if(sq_matrix(Matrix)){
-        std::cout << "Wyznacznik macierzy:" << std::endl;
-        showMatrix(Matrix, Matrix.size(), Matrix[0].size());
-        std::cout << "Wynosi: " << determinant(Matrix) << std::endl;
-    }else{
-        std::cout << "Nie można policzyć wyznacznika ponieważ macierz nie jest macierzą kwadratową" << std::endl;
-    }
-}
-
-//menu funkcji macierz
-void matrix() {
-    matrix:
-    std::cout <<"Wybierz dzialanie programu:\n+ dodawanie macierzy\n- odejmowanie macierzy\n* mnozenie macierzy\n| wyznacznik macierzy" << std::endl;
-
-    char mark;
-    std::cin >> mark;
-    switch(mark){
-        case '+':
-            addMatrix_user();
-            break;
-        case '-':
-            substractMatrix_user();
-            break;
-        case '*':
-            multiMatrix_user();
-            break;
-        case '|':
-            determinat_str();
-            break;
-        default:
-            std::cout << "Nie znana komenda \nspróbuj jeszcze raz" << std::endl;
-            goto matrix;
-    }
-}
-
-void cykl_o_tryg() {
-    std::string choose;
-    std::cout << "Chcesz skorzystać z funkcji cyklometrycznej czy trygonometrycznej" << std::endl;
-    std::cin >> choose;
-    if (choose.find("cyklome") != std::string::npos) {
-        cykl();
-    } else if (choose.find("trygono") != std::string::npos) {
-        tryg();
-    } else if (choose == "back"){
-
-    }else{
-        std::cout << "Funkcja nie znana spróbuj jeszcze raz" << std::endl;
-        cykl_o_tryg();
-    }
-}
-
-void tryg() {
-    std::cout << "Podaj funkcje i liczbe" << std::endl;
-
-    double num;
-
-    std::string fun_tryg;
-
-    std::cin >> fun_tryg >> num;
-
-    std::string wyb;
-
-    std::cout << "Wartość podajesz w radjanach czy stopniach?" << std::endl;
-
-    std::cin >> wyb;
-
-    if (wyb.find("stop") != std::string::npos) {
-        num = num / 57.3;
-    } else if (wyb.find("stop") == std::string::npos && wyb.find("radia") == std::string::npos) {
-        std::cout << "Nieznana komenda spróbuj jeszcze raz" << std::endl;
-        tryg();
-    }
-
-
-    if (fun_tryg.find("sin") != std::string::npos) {
-        std::cout << "Warotść sinusa podanego kąta wynosi " << sin(num) << std::endl;
-    } else if (fun_tryg.find("cos") != std::string::npos) {
-        std::cout << "Warotść cosinsusa podanego kąta wynosi " << cos(num) << std::endl;
-    } else if (fun_tryg.find("tan") != std::string::npos) {
-        std::cout << "Warotść tangensa podanego kąta wynosi " << tan(num) << std::endl;
-    } else if (fun_tryg.find("ctag") != std::string::npos) {
-        std::cout << "Warotść cotangensa podanego kąta wynosi " << 1 / tan(num) << std::endl;
-    } else if (fun_tryg == "back"){
-
-    } else {
-        std::cout << "Nieznane polecenie, spróbuj jeszcze raz" << std::endl;
-        tryg();
-    }
-}
-
-void cykl() {
-    std::cout << "Podaj funkcje i liczbe" << std::endl;
-
-    double num;
-
-    std::string fun_cyk;
-
-    std::cin >> fun_cyk >> num;
-
-    if (fun_cyk.find("sin") != std::string::npos) {
-        std::cout << "Warotść arcusasinusa podanego kąta wynosi " << std::fixed << std::setprecision(1)
-                  << asin(num) * 57.3 << "°" << std::endl;
-    } else if (fun_cyk.find("cos") != std::string::npos) {
-        std::cout << "Warotść arcusacosinusa podanego kąta wynosi " << std::fixed << std::setprecision(1)
-                  << acos(num) * 57.3 << "°" << std::endl;
-    } else if (fun_cyk.find("tan") != std::string::npos) {
-        std::cout << "Warotść arcusatangensa podanego kąta wynosi " << std::fixed << std::setprecision(1)
-                  << atan(num) * 57.3 << "°" << std::endl;
-    } else if (fun_cyk.find("ctag") != std::string::npos) {
-        std::cout << "Warotść arcusacotangensa podanego kąta wynosi " << std::fixed << std::setprecision(1)
-                  << M_PI / 2 - atan(num) * 57.3 << "°" << std::endl;
-    } else if (fun_cyk == "back"){
-
-    } else {
-        std::cout << "Nieznane polecenie, spróbuj jeszcze raz" << std::endl;
-        cykl();
-    }
-}
-
-void print_zao(){
-    double num;
-    int m_p_p;
-    std::cin >> num >> m_p_p;
-    std::cout << "Liczba po zaokrągleniu wynosi: " << std::fixed << std::setprecision(m_p_p) << num << std::endl;
-}
-
-double significant_figure(double n, int sig_digits){
-    if (n == 0)
-        return 0;
-    double exponent = std::floor(log10(abs(n)));
-    double multiplier = pow(10, (sig_digits - 1 - exponent));
-    return std::round(n * multiplier) / multiplier;
-}
-
-void kwadratowa(double a, double b, double c){
-    if(a == 0){
-        std::cout << "Wspolczynnik przy x^2 nie moze byc zerem"<< std::endl;
-    }
-    double delta = b * b - 4 * a * c;
-    if(delta > 0){
-        double x1 = (sqrt(delta) - b) / (2 * a);
-        double x2 = (- sqrt(delta) - b) / (2 * a);
-        std::cout<< "x1 = "<< x1 << " x2 = "<< x2 << std::endl;
-    } else if (delta == 0){
-        double x = (- b) / (2 * a);
-        std::cout << "Podwojne miejsce zerowe x0 = "<< x << std::endl;
-    } else{
-        std::cout<< "Brak rzeczywistych pierwiastkow, ale sa zespolone" << std::endl;
-        double pierwiastek_z_delty_zespolony = sqrt(sqrt(delta * delta));
-        std::cout << "x1 = " << -b/(2 * a) << " + " << pierwiastek_z_delty_zespolony/(2 * a) << "i, x2 = " << -b/(2 * a) << " - " << pierwiastek_z_delty_zespolony/(2 * a) << "i" << std::endl;
-    }
-}
-
-void kwadratowa(Complex a, Complex b, Complex c){
-    if (a.cpx_lg() == 0){
-        std::cout << "Wspolczynnik a nie moze byc zerem"<< std::endl;
-    } else {
-        Complex four(4, 0); // 4 jako liczba zespolona
-        Complex two(2, 0); // 2 jako liczba zespolona
-        //liczenie delty
-        Complex delta = (b * b) - (four * a * c);
-        Complex sqrt_delta = sqrt(delta);
-        // Oblicz pierwiastki równania
-        Complex minus_b = inerface_complex(b);
-        Complex x1 = (minus_b + sqrt_delta) / (two * a);
-        Complex x2 = (minus_b - sqrt_delta) / (two * a);
-        std::cout << "x1 = " << x1.Re << " + " << x1.Im << "i";
-        std::cout << " x2 = " << x2.Re << " + " << x2.Im << "i" << std::endl;
-    }
-}
-
-std::vector<double> solveCubic(double a, double b, double c, double d) {
-    std::vector<double> roots;
-
-    if (a == 0) {
-        throw std::invalid_argument("To nie jest rownanie trzeciego stopnia");
-    }
-
-    // Reduce the cubic equation to the form t^3 + pt + q = 0
-    double p = (3 * a * c - b * b) / (3 * a * a);
-    double q = (2 * b * b * b - 9 * a * b * c + 27 * a * a * d) / (27 * a * a * a);
-
-    // Calculate the discriminant
-    double discriminant = pow(q / 2, 2) + pow(p / 3, 3);
-
-    if (discriminant > 0) {
-        // One real root and two complex conjugate roots
-        double u = cbrt(-q / 2 + sqrt(discriminant));
-        double v = cbrt(-q / 2 - sqrt(discriminant));
-        roots.push_back(u + v - b / (3 * a));
-    } else if (discriminant == 0) {
-        // Three real roots, at least two are equal
-        double u = cbrt(-q / 2);
-        roots.push_back(2 * u - b / (3 * a));
-        roots.push_back(-u - b / (3 * a));
-    } else {
-        // Three distinct real roots
-        double r = sqrt(-pow(p / 3, 3));
-        double phi = acos(-q / (2 * r));
-        double r_cbrt = cbrt(r);
-
-        roots.push_back(2 * r_cbrt * cos(phi / 3) - b / (3 * a));
-        roots.push_back(2 * r_cbrt * cos((phi + 2 * M_PI) / 3) - b / (3 * a));
-        roots.push_back(2 * r_cbrt * cos((phi + 4 * M_PI) / 3) - b / (3 * a));
-    }
-
-    return roots;
-}
-
-
-double calka_liniowa(double a, double b, double x1, double x2){
-    double dx = 0.0001;
-    double wynik = 0;
-    if (x1 > x2){
-        return -1;
-    } else if (x1 == x2){
-        return 0;
-    } else{
-        while (x1 < x2){
-            wynik += (x1 * a + b) * dx;
-            x1 += dx;
-        }
-        return wynik;
-    }
-}
-
-double calka_kwadratowa(double a, double b, double c, double x1, double x2){
-    double dx = 0.0001;
-    double wynik = 0;
-    if (x1 > x2){
-        return -1;
-    } else if (x1 == x2){
-        return 0;
-    } else{
-        while (x1 < x2){
-            wynik += (pow(x1, 2) * a + x1 * b + c) * dx;
-            x1 += dx;
-        }
-        return wynik;
-    }
-}
-
-double calka_sin(double x1, double x2){
-    double dx = 0.0001;
-    double wynik = 0;
-    if (x1 > x2){
-        return -1;
-    } else if (x1 == x2){
-        return 0;
-    } else{
-        while (x1 < x2){
-            wynik += sin(x1) * dx;
-            x1 += dx;
-        }
-        return wynik;
-    }
-}
-
-double calka_cos(double x1, double x2){
-    double dx = 0.0001;
-    double wynik = 0;
-    if (x1 > x2){
-        return -1;
-    } else if (x1 == x2){
-        return 0;
-    } else{
-        while (x1 < x2){
-            wynik += cos(x1) * dx;
-            x1 += dx;
-        }
-        return wynik;
-    }
-}
-
-
-//obsługa całek:
-void calki() {
-    calki:
-        std::cout << "Wybierz rodzaj calki:\nL - calka z funkcji liniowej\nK - calka z funkcji kwadratowej\nS - calka z sinusa\nC - calka z cosinusa" << std::endl;
-    char mark;
-    std::cin >> mark;
-    switch (std::tolower(mark)) {
-        case 'l': {
-            double a, b, x1, x2;
-            std::cout << "Podaj kolejno: wspolczynnik a, wspolczynnik b, poczatek oraz koniec przedzialu calkowania" << std::endl;
-            std::cin >> a >> b >> x1 >> x2;
-            std::cout << "Wynik: " << calka_liniowa(a, b, x1, x2) << std::endl;
-            break;
-        }
-        case 'k': {
-            double a, b, c, x1, x2;
-            std::cout << "Podaj kolejno: wspolczynnik a, wspolczynnik b, wspolczynnik c, poczatek oraz koniec przedzialu calkowania" << std::endl;
-            std::cin >> a >> b >> c >> x1 >> x2;
-            std::cout << "Wynik: " << calka_kwadratowa(a, b, c, x1, x2) << std::endl;
-            break;
-        }
-        case 's': {
-            double x1, x2;
-            std::cout << "Podaj kolejno: poczatek oraz koniec przedzialu calkowania" << std::endl;
-            std::cin >> x1 >> x2;
-            std::cout << "Wynik: " << calka_sin(x1, x2) << std::endl;
-            break;
-        }
-        case 'c': {
-            double x1, x2;
-            std::cout << "Podaj kolejno: poczatek oraz koniec przedzialu calkowania" << std::endl;
-            std::cin >> x1 >> x2;
-            std::cout << "Wynik: " << calka_cos(x1, x2) << std::endl;
-            break;
-        }
-        default:
-            std::cout << "Wybrano nieprawidlowa operacje" << std::endl;
-            goto calki;
-    }
-}
-
-double Complex::clasic_to_tryg() const {
-    if (Im >= 0 && cpx_lg() != 0) {
-        return acos(Re / cpx_lg());
-    } else if (Im < 0) {
-        return -acos(Re / cpx_lg());
-    } else if (cpx_lg() == 0) {
-        return 7312;
-    }
-    return 7312;
-}
-
-Complex inerface_complex(Complex& cpx){
-    Complex cpx_int(cpx.Re, -cpx.Im);
-    return cpx_int;
-}
-
-void print_complex(Complex cpx){
-    if (cpx.Phi != 7321) {
-        std::cout << std::fixed << std::setprecision(3) << cpx.Re << " + i" << cpx.Im << "\ncos(" << cpx.Phi << ") + isin(" << cpx.Phi << ")" << std::endl;
-    } else
-        std::cout << "Niezdefiniowano";
-}
-
-Complex sqrt(Complex& cpx){
-    std::complex<double> cpx_pr = std::sqrt(std::complex<double>(cpx.Re, cpx.Im));
-    Complex cpx_1(real(cpx_pr), imag(cpx_pr));
-    return cpx_1;
-}
-
-void complex_choice(){
-    char oper;
-    std::vector<double> tym_kont;
-    std::cout << "Możliwe operacje do wykonania:\n1. Przekształcenie liczby zespolonej z postaci trygonometrycznej do postaci kanoniczej i na odwrót\n"
-                 "2. Dodawanie liczb zespolonych\n3. Odejmowanie liczb zespolonych\n4. Mnożenie liczb zespolonych\n5. Dzielenie liczb zespolonych\n"
-                 "6. Znajdowanie liczby sprzeżonej\n 7.Obliczanie długości liczby zespolonej\n\nKtórą z operacji chcesz wykonać?" << std::endl;
-    std::cin >> oper;
-    if (oper == 1){
-        double re;
-        double im;
-        std::cout << "Podaj liczbe zespoloną" << std::endl;
-        std::cin >> re;
-        std::cin >> im;
-        if (im != 0) {
-            Complex zes(re, im);
-            print_complex(zes);
-        } else {
-            re = re * M_PI / 180;
-            Complex zes(re);
-            print_complex(zes);
-        }
-    } else if (oper == 2 || oper == 3 || oper == 4 || oper == 5){
-        std::cout << "Podaj liczby zespolone" << std::endl;
-        double re_1, im_1, re_2, im_2;
-        std::cin >> re_1;
-        std::cin >> im_1;
-        std::cin >> re_2;
-        std::cin >> im_2;
-        Complex comp_1(re_1, im_1);
-        Complex comp_2(re_2, im_1);
-        if (oper == 2){
-            print_complex(comp_1 + comp_2);
-        } else if (oper == 3){
-            print_complex(comp_1 - comp_2);
-        } else if (oper == 4){
-            print_complex(comp_1 * comp_2);
-        } else if (oper == 5){
-            print_complex(comp_1 / comp_2);
-        }
-    } else if (oper == 6 || oper == 7){
-        double re;
-        double im;
-        std::cout << "Podaj liczby zespolone" << std::endl;
-        std::cin >> re;
-        std::cin >> im;
-        Complex comple(re, im);
-        if (oper == 6){
-            print_complex(inerface_complex(comple));
-        } else{
-            std::cout << comple.cpx_lg();
-        }
-    } else {
-        char wyb;
-        std::cout << "Nie zdefiniowano, chcesz wrócić do menu [T/N]" << std::endl;
-        std::cin >> wyb;
-        if (wyb == 'N'){
-            complex_choice();
-        }
-    }
-}
-
-bool does_triangle_exist(double a, double b, double c){
-    if (a + b > c) return true;
-    return false;
-}
-
-double triangleArea1(double a, double h){
-    return a * h / 2;
-}
-
-double triangleArea2(double a, double b, double c){
-    double p = (a + b + c) / 2;
-    double area = sqrt(p * (p - a) * (p - b) * (p - c));
-    return significant_figure(area, 8);
-}
-
-double triangleArea3(double a, double b, double angle){
-    double area = a * b * sin(angle*3.14159/180) / 2;
-    return significant_figure(area, 8);
-}
-
-void triangle(){
-    triangle:
-    std::cout << "Wybierz jakim sposobem chcialbys obliczyc pole trojkata\n1 - dlugosc boku i wysokosci na niej opuszczonej\n"
-                 "2 - znane sa wszystkie dlugosci bokow trojkata\n3 - posiadamy dlugosci dwoch bokow i miare kata pomiedzy nimi" << std::endl;
-    int mark;
-    std::cout << "Wybierz cyfre:\n";
-    std::cin >> mark;
-
-    double a, b, c, h, angle;
-    double area;
-
-    switch( mark ){
-        case 1:
-            std::cout << "Obliczanie pola przez podanie boku i wysokosci." << std::endl;
-        bok11:
-            std::cout << "Dlugosc boku:\n";
-            std::cin >> a;
-            if (a <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok11;
-            }
-        wysokosc11:
-            std::cout << "Dlugosc wysokosci:\n";
-            std::cin >> h;
-            if (h <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto wysokosc11;
-            }
-            area = triangleArea1(a, h);
-            break;
-
-        case 2:
-        triangle2:
-            std::cout << "Obliczanie pola przez podanie dlugosci wszystkich bokow." << std::endl;
-        bok21:
-            std::cout << "Dlugosc pierwszego boku:\n";
-            std::cin >> a;
-            if (a <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok21;
-            }
-        bok22:
-            std::cout << "Dlugosc drugiego boku:\n";
-            std::cin >> b;
-            if (b <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok22;
-            }
-        bok23:
-            std::cout << "Dlugosc trzeciego boku:\n";
-            std::cin >> c;
-            if (c <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok23;
-            }
-
-            if(does_triangle_exist(a,b,c) && does_triangle_exist(c,a,b) && does_triangle_exist(b,c,a)) {
-                area = triangleArea2(a, b, c);
-            }
-            else {
-                std::cout << "Podano nieprawidlowe dlugosci w trojkacie. Trojkat nie istnieje.\n"
-                             "Sprobuj jeszcze raz" << std::endl;
-                goto triangle2;
-            }
-            break;
-
-        case 3:
-            std::cout << "Obliczanie pola przez podanie dlugosci dwoch bokow i kata pomiedzy nimi." << std::endl;
-        bok31:
-            std::cout << "Dlugosc pierwszego boku:\n";
-            std::cin >> a;
-            if (a <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok31;
-            }
-        bok32:
-            std::cout << "Dlugosc drugiego boku:\n";
-            std::cin >> b;
-            if (b <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok32;
-            }
-        miara:
-            std::cout << "Miara kata:\n";
-            std::cin >> angle;
-            if ((angle <= 0) || (angle > 180)){
-                std::cout << "Podano nieprawidlowa miare. Sprobuj ponownie." << std::endl;
-                goto miara;
-            }
-
-            area = triangleArea3(a, b, angle);
-            break;
-        default:
-            std::cout << "Wybrano nieprawidlowa operacje" << std::endl;
-            //blad();
-            goto triangle;
-    }
-
-    std::cout << "Pole trojkata wynosi: " << area << std::endl;
-}
-
-double rectangularArea(double a, double b){
-    return a * b;
-}
-
-void rectangular(){
-    double a;
-    double b;
-    double area;
-    std::cout << "Podaj dlugosci bokow prostokata: " << std::endl;
-    bok1:
-    std::cout << "Dlugosc pierwszego boku:\n";
-    std::cin >> a;
-    if (a <= 0 ){
-        std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-        goto bok1;
-    }
-    bok2:
-    std::cout << "Dlugosc drugiego boku:\n";
-    std::cin >> b;
-    if (b <= 0 ){
-        std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-        goto bok2;
-    }
-    area = rectangularArea(a, b);
-    std::cout << "Pole prostokatu wynosi: " << area << std::endl;
-}
-
-double diamondArea1(double a, double h){
-    return a * h;
-}
-
-double diamondArea2(double e, double f){
-    return e * f / 2;
-}
-
-void diamond(){
-    diamond:
-    std::cout << "Wybierz jakim sposobem chcialbys obliczyc pole trapezu\n1 - dlugosc boku i wysokosci na niej opuszczonej\n"
-                 "2 - znane sa dlugosci przekatnych rombu" << std::endl;
-
-    int mark;
-    std::cout << "Wybierz cyfre:\n";
-    std::cin >> mark;
-
-    double a, h;
-    double area;
-
-    switch( mark ){
-        case 1:
-            std::cout << "Obliczanie pola przez podanie boku i wysokosci." << std::endl;
-        bok11:
-            std::cout << "Dlugosc boku:\n";
-            std::cin >> a;
-            if (a <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok11;
-            }
-        wysokosc11:
-            std::cout << "Dlugosc wysokosci:\n";
-            std::cin >> h;
-            if (h <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto wysokosc11;
-            }
-            area = diamondArea1(a, h);
-            break;
-
-        case 2:
-            std::cout << "Obliczanie pola przez podanie dlugosci przekatnych." << std::endl;
-        bok21:
-            std::cout << "Dlugosc pierwszej przekatnej:\n";
-            std::cin >> a;
-            if (a <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok21;
-            }
-        bok22:
-            std::cout << "Dlugosc drugiej przekatnej:\n";
-            std::cin >> h;
-            if (h <= 0 ){
-                std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-                goto bok22;
-            }
-            area = diamondArea2(a, h);
-            break;
-
-        default:
-            std::cout << "Wybrano nieprawidlowa operacje" << std::endl;
-            //blad();
-            goto diamond;
-    }
-    std::cout << "Pole rombu wynosi: " << area << std::endl;
-}
-
-double trapezeArea(double a, double b, double h){
-    return (a + b) * h / 2;
-}
-
-void trapeze(){
-    double a;
-    double b;
-    double h;
-    double area;
-    std::cout << "Podaj dlugosci podstaw trapezu: " << std::endl;
-    bok1:
-    std::cout << "Dlugosc pierwszej podstawy:\n";
-    std::cin >> a;
-    if (a <= 0 ){
-        std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-        goto bok1;
-    }
-    bok2:
-    std::cout << "Dlugosc drugiej podstawy:\n";
-    std::cin >> b;
-    if (b <= 0 ){
-        std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-        goto bok2;
-    }
-    wysokosc:
-    std::cout << "Dlugosc wysokosci:\n";
-    std::cin >> h;
-    if (h <= 0 ){
-        std::cout << "Podano nieprawidlowa dlugosc. Sprobuj ponownie." << std::endl;
-        goto wysokosc;
-    }
-    area = trapezeArea(a, b, h);
-    std::cout << "Pole trapezu wynosi: " << area << std::endl;
-}
-
-void area(){
-    area:
-    std::cout <<"Wybierz dzialanie programu:\nP obliczanie pola prostokata\nT obliczanie pola trojkata\nR obliczanie pola rombu\nN obliczanie pola trapezu" << std::endl;
-
-    char mark;
-    std::cin >> mark;
-    switch( std::tolower(mark) ){
-        case 'p':
-            rectangular();
-            break;
-        case 't':
-            triangle();
-            break;
-        case 'r':
-            diamond();
-            break;
-        case 'n':
-            trapeze();
-            break;
-        default:
-            std::cout << "Wybrano nieprawidlowa operacje" << std::endl;
-            //blad();
-            goto area;
-    }
-}
-
-double logarytmy(double liczba_logarytmowana, double podstawa_logarytmu ){
-    double wynik = log(liczba_logarytmowana) / log(podstawa_logarytmu);
-    return wynik;
-}
-
-void logarytmy_by_Natalia(){
-    std::cout << "Podaj liczbe do zlogarytmowanie i podstawe logarytmu:" << std::endl;
-    double ll, base;
-    std::cin >> ll >> base;
-    std::cout << "Wynik logarytmu to" << logarytmy(ll,base);
-}
-
-//
-//void przelicznik(){
-//    std::cout << "Wybierz jaki rodzaj jednostek chcesz prezliczyć." << std::endl << "1. Pr";
-//}
-
-void prosty() {
-    prosty:
-    int wyb;
-    double x, y;
-    std::cout << "Jakie dzialanie chcesz wykonac? \n"
-                 "1. Dodawanie\n"
-                 "2. Odejmowanie\n"
-                 "3. Mnozenie\n"
-                 "4. Dzielenie\n"
-                 "5. Potegowanie\n"
-                 "6. Znalezienie reszty z dzielenia\n"
-                 "7. Wyjscie\n" << std::endl;
-    std::cin >> wyb;
-
-    switch(wyb) {
-        case 1:
-            std::cout << "Podaj liczby, ktore chcesz dodac" << std::endl;
-            std::cin >> x >> y;
-            std::cout << "Wynik: " << addition(x, y) << std::endl;
-            goto prosty;
-        case 2:
-            std::cout << "Podaj liczby, ktore chcesz odjac" << std::endl;
-            std::cin >> x >> y;
-            std::cout << "Wynik: " << subtraction(x, y) << std::endl;
-            goto prosty;
-        case 3:
-            std::cout << "Podaj liczby, ktore chcesz pomnozyc" << std::endl;
-            std::cin >> x >> y;
-            std::cout << "Wynik: " << multiplication(x, y) << std::endl;
-            goto prosty;
-        case 4:
-            std::cout << "Podaj liczby, ktore chcesz podzielic" << std::endl;
-            std::cin >> x >> y;
-            if (y == 0) {
-                std::cout << "Blad: dzielenie przez zero!" << std::endl;
-            } else {
-                std::cout << "Wynik: " << division(x, y) << std::endl;
-            }
-            goto prosty;
-        case 5:
-            std::cout << "Podaj podstawe i wykladnik" << std::endl;
-            std::cin >> x >> y;
-            std::cout << "Wynik: " << exponential(x, y) << std::endl;
-            goto prosty;
-        case 6:
-            int a, b;
-            std::cout << "Podaj liczby, ktore chcesz podzielic i obliczyc reszte" << std::endl;
-            std::cin >> a >> b;
-            if (b == 0) {
-                std::cout << "Blad: dzielenie przez zero!" << std::endl;
-            } else {
-                std::cout << "Wynik: " << modulo(a, b) << std::endl;
-            }
-            goto prosty;
-        case 7:
-            break;
-        default:
-            std::cout << "Nieprawidlowy wybor. Sprobuj ponownie." << std::endl;
-            goto prosty;
-    }
-}
-
